refactor: 국가 코드 매핑 상수 분리
