+++ apps/admin/src/components/features/univ-apply-infos/countryCodeAliases.ts
@@ -1,90 +1,4 @@
-export const COUNTRY_CODE_BY_NAME: Record<string, string> = {
-	// 한국어
-	오스트리아: "AT",
-	호주: "AU",
-	아제르바이잔: "AZ",
-	벨기에: "BE",
-	브루나이: "BN",
-	브라질: "BR",
-	캐나다: "CA",
-	콜롬비아: "CO",
-	스위스: "CH",
-	중국: "CN",
-	체코: "CZ",
-	에스토니아: "EE",
-	독일: "DE",
-	덴마크: "DK",
-	스페인: "ES",
-	핀란드: "FI",
-	프랑스: "FR",
-	영국: "GB",
-	아일랜드: "IE",
-	홍콩: "HK",
-	헝가리: "HU",
-	인도네시아: "ID",
-	이스라엘: "IL",
-	이탈리아: "IT",
-	일본: "JP",
-	카자흐스탄: "KZ",
-	리투아니아: "LT",
-	말레이시아: "MY",
-	네덜란드: "NL",
-	노르웨이: "NO",
-	포르투갈: "PT",
-	러시아: "RU",
-	스웨덴: "SE",
-	싱가포르: "SG",
-	태국: "TH",
-	튀르키예: "TR",
-	대만: "TW",
-	미국: "US",
-	우즈베키스탄: "UZ",
-	// 영어 풀 네임
-	Austria: "AT",
-	Australia: "AU",
-	Azerbaijan: "AZ",
-	Belgium: "BE",
-	Brunei: "BN",
-	Brazil: "BR",
-	Canada: "CA",
-	Colombia: "CO",
-	Switzerland: "CH",
-	China: "CN",
-	"Czech Republic": "CZ",
-	Czechia: "CZ",
-	Czech: "CZ",
-	Estonia: "EE",
-	Germany: "DE",
-	Denmark: "DK",
-	Spain: "ES",
-	Finland: "FI",
-	France: "FR",
-	"United Kingdom": "GB",
-	UK: "GB",
-	"Hong Kong": "HK",
-	Hungary: "HU",
-	Ireland: "IE",
-	Indonesia: "ID",
-	Israel: "IL",
-	Italy: "IT",
-	Japan: "JP",
-	Kazakhstan: "KZ",
-	Lithuania: "LT",
-	Malaysia: "MY",
-	Netherlands: "NL",
-	Norway: "NO",
-	Portugal: "PT",
-	Russia: "RU",
-	Sweden: "SE",
-	Singapore: "SG",
-	Thailand: "TH",
-	Turkey: "TR",
-	Türkiye: "TR",
-	Taiwan: "TW",
-	"United States": "US",
-	USA: "US",
-	Uzbekistan: "UZ",
-};
+import { COUNTRY_CODE_BY_NAME } from "./countryCodeConstants";
 
 export function resolveCountryCode(value: string): string {
 	return COUNTRY_CODE_BY_NAME[value.trim()] ?? value;
